+++ src/wrench/context.py
@@ -15,11 +15,33 @@
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
 
+import functools
 from typing import Callable, Dict, Iterable
 
 from requests_gpgauthlib import GPGAuthSession
 
 from .models import Group, User
+
+
+def cached(func):
+    """
+    Store the result of the function call in the object `__cache` attribute so that subsequent calls to the function
+    with the same parameters just return the cached value.
+    """
+    @functools.wraps(func)
+    def inner(*args, **kwargs):
+        obj = args[0]
+
+        if not hasattr(obj, '__cache'):
+            obj.__cache = {}
+
+        cache_key = (func, args[1:], frozenset(kwargs.items()))
+        if cache_key not in obj.__cache:
+            obj.__cache[cache_key] = func(*args, **kwargs)
+
+        return obj.__cache[cache_key]
+
+    return inner
 
 
 class Context:
@@ -42,43 +64,54 @@ class Context:
         self._groups_by_id = None  # type: Dict[str, Group]
 
     @property  # type: ignore
+    @cached
     def users(self) -> Iterable[User]:
         """
         Return an iterable of :class:`User` objects from Passbolt. The result is cached so calling this method multiple
         times will trigger only one request to Passbolt.
         """
-        if self._users is None:
-            self._users = self.get_users_func(self.session)
+        return self.get_users_func(self.session)
 
-        return self._users
+    def get_users_by(self, attr):
+        return {getattr(user, attr): user for user in self.users}
 
     @property  # type: ignore
+    @cached
     def users_by_id(self) -> Dict[str, User]:
         """
         Return a dict in the form `{id: user}` of the Passbolt users.
         """
-        if self._users_by_id is None:
-            self._users_by_id = {user.id: user for user in self.users}
-
-        return self._users_by_id
+        return {user.id: user for user in self.users}
 
     @property  # type: ignore
+    @cached
+    def users_by_name(self) -> Dict[str, User]:
+        """
+        Return a dict in the form `{username: user}` of the Passbolt users.
+        """
+        return {user.username: user for user in self.users}
+
+    @property  # type: ignore
+    @cached
     def groups(self) -> Iterable[Group]:
         """
         Return an iterable of :class:`Group` objects from Passbolt. The result is cached so calling this method multiple
         times will trigger only one request to Passbolt.
         """
-        if self._groups is None:
-            self._groups = self.get_groups_func(self.session)
-
-        return self._groups
+        return self.get_groups_func(self.session)
 
     @property  # type: ignore
+    @cached
     def groups_by_id(self) -> Dict[str, Group]:
         """
         Return a dict in the form `{id: group}` of the Passbolt groups.
         """
-        if self._groups_by_id is None:
-            self._groups_by_id = {group.id: group for group in self.groups}
+        return {group.id: group for group in self.groups}
 
-        return self._groups_by_id
+    @property  # type: ignore
+    @cached
+    def groups_by_name(self) -> Dict[str, Group]:
+        """
+        Return a dict in the form `{group_name: group}` of the Passbolt groups.
+        """
+        return {group.name: group for group in self.groups}